+++ klarsatz/__init__.py
@@ -4,6 +4,6 @@ from .interpreter import Interpreter
 from .lexer import lexer, norm
 from .parser import Parser
 
-__version__ = "0.8.0"
+__version__ = "0.8.1"
 __all__ = ["Interpreter", "KlarsatzFehler", "LaufzeitFehler", "LimitFehler", "SyntaxFehler",
            "Parser", "lexer", "norm", "__version__"]